Modifications in the frontend of bikes

+++ routes/web.php
@@ -25,11 +25,11 @@ Route::middleware(['auth.role:admin'])->group(function () {
     Route::post('/admin/part/save', 'App\Http\Controllers\Admin\PartController@save')->name('admin.part.save');
     Route::get('/admin/part/show/{id}', 'App\Http\Controllers\Admin\PartController@show')->name('admin.part.show');
     Route::delete('/admin/part/remove/{id}', 'App\Http\Controllers\Admin\PartController@remove')->name('admin.part.remove');
-    Route::get('/admin/showAll', 'App\Http\Controllers\Admin\BikeController@showAll')->name("admin.bike.showAll");
-    Route::get('/admin/create', 'App\Http\Controllers\Admin\BikeController@create')->name("admin.bike.create");
-    Route::post('/admin/save', 'App\Http\Controllers\Admin\BikeController@save')->name("admin.bike.save");
-    Route::get('/admin/show/{id}', 'App\Http\Controllers\Admin\BikeController@show')->name("admin.bike.show");
-    Route::delete('/remove/{id}', 'App\Http\Controllers\Admin\BikeController@remove')->name("admin.bike.remove");
+    Route::get('/admin/bike/', 'App\Http\Controllers\Admin\BikeController@showAll')->name("admin.bike.showall");
+    Route::get('/admin/bike/create', 'App\Http\Controllers\Admin\BikeController@create')->name("admin.bike.create");
+    Route::post('/admin/bike/save', 'App\Http\Controllers\Admin\BikeController@save')->name("admin.bike.save");
+    Route::get('/admin/bike/show/{id}', 'App\Http\Controllers\Admin\BikeController@show')->name("admin.bike.show");
+    Route::delete('admin/bike/remove/{id}', 'App\Http\Controllers\Admin\BikeController@remove')->name("admin.bike.remove");
 });
 
 //User routes
